app/requirements.txt
requests==2.31.0
bs4==0.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
